mlwh_connection: avoid requiring the DB when importing the app.
Tries to solve the issue of needing the DB_URL env var
being set even when testing. This is especially confusing as
it will not be respected: when running pytest, this connection
is overriden and a URL defined in a ini file is used.
So here the env var (and the engine) are only created when first used,
rather than at import time (this file is imported by the main app module).

Signed-off-by: Adam Blanchet <ab59@sanger.ac.uk>