Fix 404s on loading dynamic vue routes.
When serving the static built Vue app, there is no nodeJS server
to do the routing dynamically. The vue router would not be there
to intercept the request and serve the appropriate page. Instead,
vue-router's documentation suggests to make the web server fallback
to serving /index.html, and then the SPA will render the appropriate
page.

See https://router.vuejs.org/guide/essentials/history-mode.html#apache

It might be useful to add a 404 handler in the Vue app, see
https://router.vuejs.org/guide/essentials/history-mode.html#caveat

Signed-off-by: Adam Blanchet <ab59@sanger.ac.uk>