Disable FallbackResource directive for the assets folder.
We believe we can safely disable the FallbackResource directive
for the assets folder, as these are either there or not.
The browser either can fetch these static resources or cannot,
these are not web pages the browser can interpret the JS of. It
would be confusing to fetch a static image and get the fallback
index.html page served instead.

Signed-off-by: Adam Blanchet <ab59@sanger.ac.uk>