Fixed small bugs and improved API documentation.
Fixed a bug in assigning the QC state outcome to the model
that is served by FastAPI. Fixed a problem in test data,
which did not allow to correctly test this functionality.

To allow for back-filling QC database dataa added
data_updated argument to the assign_qc_state method.

Introduced constants for most of default and hardcoded
values.

Fixed a bug in assigning a default value to the
is_preliminary argument of the assign_qc_state method.

Extended teh documentation.