Reduced run time for tests.
Set the scope of fixtures to 'module' or 'package' where
possible. The fixtures that create databases are now run
once per module (test file). Sourcing of credentials for
databases is done once per the pytest run ('package'
scope).

Added the teardown section (after the 'yield' statement)
to fixtures with the 'function' scope (the default scope
of the pytest fixtures). The teardown section deletes all
data that the fixture loaded to the databases.

Dropped altering the settings of the mlwh database at the
stage of creating the schema. Instead, foreign keys
constraints are now dropped at the teardown section and
reinstated once data deletion is done.