Added a model to consolidate all LIMS data.
Added a new model lang_qc.models.pacbio.lims to wrap around
LIMS data for PacBio entities.

The model is capable of representing both a well with either a
single sample or multiple samples and a single sample (library).

This model will be used by the reimplemented endpoint for
a single well.