Removed qc_flow_status key from JSON response.
qc_flow_status key was a part of the response to a request
to fetch wells for a particular QC flow status. This key
is not being used by the front end code. Its presence
makes the response model less generic, i.e. not usable
for requests that are not associated with any particular
QC flow status.