Added local ORM classes for mlwh database.
Copied a few classes we need from the external
ml_warehouse package to lang_qc.db.mlwh_schema.
Reduced the number of imports to the types that
are actually needed. Reduced the number of mapped
columns for the study and sample tables to columns
we are or might be using.