Refactored end points for claiming and updating QC.
Endpoint URLs are changed. Product ID, but no other
identifier, is embedded into the URLs. The database
helper is changed to reflect this change. It is also
extended to accommodate an extra identifier - the
plate number. Mapping of product IDs to run name,
well label and plate number is done via the mlwh,
records therefore creating QC records for entities
that are not present in mlwh becomes impossible.