QC states retrieval refactored, bug fixed.
In future lang_qc/db/helper/qc.py will be used as a platform-
independent retrieved of QC states for a single product or lists
of products. The current designed is changed to pure functional
since no state is associated with the proposed functionality.

Static type annotations are used to help with understanding
input and output data.

All functions apart from the one used externally are converted
to private. Tests are added for the only remaining public function.

A bug (incorrect implementation for the sequencing_outcomes_only
flag) is fixed.

The module converted to pure functional