Added a method to flag products that have QC state.
Added a new method, qc_state_for_products_exists, to perform
a check for QC states for a list of product ids as a single query.
Used this new method in a calculation for the upcoming tab
instead of multiple calls to a similar methods for a single
prodict id.