Ensure empty lists of qc states are not returned.
The descripton of the 'get_qc_states_by_id_product_list' method
in npg_langqc.db.helper.qc says that when sequencing qc states
are requested "it is guaranteed that the list of QcState objects
has only one member". However, the implementation did not warrant
this. It was possible to get an empty list of qc states.