Introduced 'On hold external' QC state.
This state will be assigned to wells when the
outcome of the QC assessment depends on liasing
with a third party. The existing 'On hold' state
will be used for flagging internal investigations.

As fas as QC statuses and workflow is concerned,
the new state is handled in the same way as the
existing 'On hold' state.